refactor(devnet): restructure lib.rs to match workspace conventions
