+++ system_tests/README.md
@@ -1,0 +1,6 @@
+# `system_tests`
+
+System/E2E test infrastructure for Base node.
+
+This crate provides a programmatic way to spin up complete L1+L2 devnet stacks
+using testcontainers for container orchestration.
